+++ rollup.config.js
@@ -1,12 +1,12 @@
 import typescript from '@rollup/plugin-typescript';
-import { terser } from 'rollup-plugin-terser';
+import cleanup from 'rollup-plugin-cleanup';
 import pkg from './package.json';
 
 export default [
   {
     external: ['leac', 'peberminta'],
     input: 'src/parseley.ts',
-    plugins: [typescript(), terser()],
+    plugins: [typescript(), cleanup({ extensions: ['ts'] })],
     output: [
       { file: pkg.main, format: 'cjs' },
       { file: pkg.module, format: 'es' },